Restructure project into multi-module three-tier layout
