+++ README.md
@@ -1,9 +1,10 @@
 # codex_test
 
-This is a minimal Spring Boot application with a simple REST controller.
+This repository has been restructured into a multi‑module project representing a simple three tier architecture.
 
-### Endpoints
-- `GET /demo/{id}` – returns the provided id
-- `POST /demo` – echoes the posted body
+## Modules
+- **presentation** – entry point for the application (e.g. controllers)
+- **business** – application logic layer
+- **data** – persistence layer
 
-
+Each module is an independent Spring Boot project initialized with the standard Spring Boot structure.
