+++ business/src/main/java/com/example/business/BusinessApplication.java
@@ -1,0 +1,13 @@
+package com.example.business;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BusinessApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BusinessApplication.class, args);
+    }
+
+}
