+++ business/src/test/java/com/example/business/BusinessApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.business;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BusinessApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
