+++ data/src/main/java/com/example/data/DataApplication.java
@@ -1,0 +1,13 @@
+package com.example.data;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DataApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DataApplication.class, args);
+    }
+
+}
