+++ data/src/test/java/com/example/data/DataApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.data;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DataApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
