+++ presentation/src/main/java/com/example/presentation/PresentationApplication.java
@@ -1,11 +1,13 @@
-package com.example.demo;
+package com.example.presentation;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class DemoApplication {
+public class PresentationApplication {
+
     public static void main(String[] args) {
-        SpringApplication.run(DemoApplication.class, args);
+        SpringApplication.run(PresentationApplication.class, args);
     }
+
 }
