+++ presentation/src/test/java/com/example/presentation/PresentationApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.presentation;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PresentationApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
